+++ chat-widget/src/common/Constants.ts
@@ -206,6 +206,20 @@ export enum EnvironmentVersion {
     test = "test"
 }
 
+export enum E2VVOptions{
+    NoCalling = "192350000",
+    VideoAndVoiceCalling = "192350001",
+    VoiceOnly = "192350002"
+}
+
+export enum LiveWorkItemState  {
+    Active = "Active",
+    Closed = "Closed",
+    Open = "Open",
+    Waiting = "Waiting",
+    WrapUp = "WrapUp"
+}
+
 export class TranscriptConstants {
     public static readonly ChatTranscriptsBodyColor = "#F5F5F5";
     public static readonly TranscriptMessageEmojiMessageType = "http://schema.skype.com/emoji";