+++ chat-widget/src/components/livechatwidget/LiveChatWidget.stories.tsx
@@ -759,7 +759,7 @@ const liveChatWidgetPopoutStyleProps: ILiveChatWidgetProps = {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     chatSDK: new MockChatSDK() as any,
     controlProps: {
-        skipChatButtonRendering: true,
+        hideStartChatButton: true,
         hideHeader: true
     },
     styleProps: {
@@ -787,6 +787,9 @@ const liveChatWidgetReconnectChatPaneProps: ILiveChatWidgetProps = {
     chatConfig: {
         LiveChatConfigAuthSettings: {
             msdyn_javascriptclientfunction: "testAuth"
+        },
+        LiveWSAndLiveChatEngJoin: {
+            msdyn_enablechatreconnect: "true"
         }
     },
     styleProps: {
@@ -796,9 +799,6 @@ const liveChatWidgetReconnectChatPaneProps: ILiveChatWidgetProps = {
             top: "20px",
             left: "20px"
         }
-    },
-    reconnectChatPaneProps: {
-        isReconnectEnabled: true
     }
 };