+++ chat-widget/src/components/reconnectchatpanestateful/interfaces/IReconnectChatPaneStatefulProps.ts
@@ -1,7 +1,6 @@
 import { IReconnectChatPaneProps } from "@microsoft/omnichannel-chat-components/lib/types/components/reconnectchatpane/interfaces/IReconnectChatPaneProps";
 
 export interface IReconnectChatPaneStatefulProps extends IReconnectChatPaneProps {
-    isReconnectEnabled?: boolean;
     reconnectId?: string;
     redirectInSameWindow?: boolean;
 }