+++ chat-widget/src/components/webchatcontainerstateful/webchatcontroller/middlewares/storemiddlewares/dataMaskingMiddleware.ts
@@ -54,6 +54,7 @@ const applyDataMasking = (action: IWebChatAction, regexCollection: IDataMaskingR
                 });
             }
         }
+
         // Exit if rule matched
         if (isRuleMatched === true) {
             break;