+++ CHANGE_LOG.md
@@ -65,4 +65,17 @@ All notable changes to this project will be documented in this file.## [Unreleas
 - Fixed multiple scrollbars in PostChat to display single scroll bar.
 - Fixed chat loading issue, Updating datamasking events
 - Added Customer Voice Form Event loggin for PostChat Survey (V1 Parity)
+- Removed isReconnectEnabled from props
 - Fixed OutOfOffice Chat Button Icon appearance
+- Adding support for customizing widget scroll bar
+- Clearing live chat context for raise error event
+- Renaming `skipChatButtonRendering` to `hideStartChatButton`
+- Refactoring reconnect functionality code
+- Fixing chat widget refresh bug showing reconnect pane
+- Supporting to start new chat if localStorage has stale Active state
+- Fixing end chat to gracefully complete if auth token has expired
+- Refactoring input validation pane and replacing keyDown event with keyUp
+- Fixed download and email transcript after post chat survey is sent / rendered
+- Adding ESC shortcut for confirmation dismiss, Disable sendBox for linked survey
+- Fixed styles for input validation pane
+- Uptake `chat-components` 0.1.0-main.a7ac5de